add i18n; impl auto db cleanup

+++ src/main/java/tech/makcymal/polylang/talks/transcription/Task.java
@@ -8,13 +8,13 @@ import java.util.UUID;
 public class Task {
 
     private UUID talkId;
-    private int chunkStart;
+    private float chunkStart;
     private String fileToTranscribe;
     private String fileToProcess;
     private Throwable transcribingError;
     private Throwable processingError;
 
-    public Task(UUID talkId, int chunkStart, String fileToTranscribe, String fileToProcess) {
+    public Task(UUID talkId, float chunkStart, String fileToTranscribe, String fileToProcess) {
         this.talkId = talkId;
         this.chunkStart = chunkStart;
         this.fileToTranscribe = fileToTranscribe;
